backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.22,<1.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0,<2.0
passlib[bcrypt]>=1.7.4,<2.0
python-jose[cryptography]>=3.3.0,<4.0
pydantic>=2.6,<3.0
email-validator>=2.1,<3.0
python-multipart>=0.0.9,<0.1
alembic>=1.13,<2.0
boto3>=1.34,<2.0